requirements.txt
Pillow==10.2.0
numpy==1.26.4
pyinstaller==6.3.0